chore(design-system,message-list): move FavouriteButtonIcon out from design system and update its colour
